feat(item): update item details
